backend/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain_pinecone
sentence-transformers
pinecone-client
langchain-google-genai
python-dotenv
pypdf

